08-08 병합

+++ src/main/java/com/green/beadalyo/jhw/user/UserServiceImpl.java
@@ -98,6 +98,9 @@ public class UserServiceImpl implements UserService{
     public void deleteProfileImage() {
         User user = repository.getReferenceById(authenticationFacade.getLoginUserPk());
         try {
+            if(user.getUserPic() == null) {
+                return;
+            }
             String delAbsoluteFolderPath = String.format("%s", customFileUtils.uploadPath);
             File file = new File(delAbsoluteFolderPath, user.getUserPic());
             file.delete();
